+++ CHANGES.rst
@@ -4,7 +4,8 @@ Changelog
 0.2 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- Also lookup behaviors when searching for local role fields.
+  [cedricmessiant]
 
 
 0.1 (2014-10-24)